+++ board/rcm/common/flw_sf.h
@@ -12,11 +12,11 @@
 
 int flw_spi_flash_found(unsigned int bus, unsigned int cs);
 
-int flw_spi_flash_erase(struct flw_dev_t* fd, unsigned int addr, unsigned int size);
+int flw_spi_flash_erase(struct flw_dev_t* fd, unsigned long addr, unsigned long size);
 
-int flw_spi_flash_write(struct flw_dev_t* fd, unsigned int addr, unsigned int size, const char* data);
+int flw_spi_flash_write(struct flw_dev_t* fd, unsigned long addr, unsigned long size, const char* data);
 
-int flw_spi_flash_read(struct flw_dev_t* fd, unsigned int addr, unsigned int size, char* data);
+int flw_spi_flash_read(struct flw_dev_t* fd, unsigned long addr, unsigned long size, char* data);
 
 void flw_spi_flash_list_add(void);
